+++ src/store/path/getPath.js
@@ -1,0 +1,3 @@
+import { store } from '../store.js';
+
+export const getPath = () => store.currentPath;
