+++ src/store/path/updatePath.js
@@ -1,0 +1,3 @@
+import { store } from '../store.js';
+
+export const updatePath = (newPath) => (store.currentPath = newPath);
