+++ src/store/readLineInterface/getReadLineInterface.js
@@ -1,0 +1,3 @@
+import { store } from '../store.js';
+
+export const getReadLineInterface = () => store.readlineInterface;
