+++ src/store/store.js
@@ -1,0 +1,7 @@
+// import { stdin as input, stdout as output } from 'node:process';
+
+export const store = {
+  currentPath: null,
+  userName: null,
+  readlineInterface: null,
+};
