+++ src/store/userName/getUserName.js
@@ -1,0 +1,3 @@
+import { store } from '../store.js';
+
+export const getUserName = () => store.userName;
