+++ src/store/userName/saveUserName.js
@@ -1,0 +1,3 @@
+import { store } from '../store.js';
+
+export const saveUserName = (userName) => (store.userName = userName);
